+++ rs/ledger_suite/icrc1/test_utils/src/lib.rs
@@ -35,6 +35,8 @@ use std::sync::Arc;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 use strum::EnumCount;
 
+pub mod icrc3;
+
 pub const E8: u64 = 100_000_000;
 pub const DEFAULT_TRANSFER_FEE: u64 = 10_000;