+++ rs/sns/governance/src/governance/test_helpers.rs
@@ -1,6 +1,7 @@
 use super::*;
 use crate::pb::v1::{Motion, NeuronPermissionType};
 use async_trait::async_trait;
+use candid::Nat;
 use ic_nervous_system_clients::canister_status::{
     CanisterStatusResultFromManagementCanister, CanisterStatusResultV2, CanisterStatusType,
 };
@@ -122,6 +123,19 @@ impl ICRC1Ledger for DoNothingLedger {
 
     fn canister_id(&self) -> CanisterId {
         CanisterId::from(42)
+    }
+
+    async fn icrc2_approve(
+        &self,
+        _spender: Account,
+        _amount: u64,
+        _expires_at: Option<u64>,
+        _fee: u64,
+        _from_subaccount: Option<Subaccount>,
+    ) -> Result<Nat, NervousSystemError> {
+        Err(NervousSystemError {
+            error_message: "Not Implemented".to_string(),
+        })
     }
 
     async fn icrc3_get_blocks(